refactor: add unit-selectcommunity module
