+++ core-commonui/unit-selectcommunity/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/selectcommunity/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.selectcommunity.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.selectcommunity.SelectCommunityMviModel
+import org.koin.java.KoinJavaComponent
+
+actual fun getSelectCommunityViewModel(): SelectCommunityMviModel {
+    val res: SelectCommunityMviModel by KoinJavaComponent.inject(SelectCommunityMviModel::class.java)
+    return res
+}
