+++ core-commonui/unit-selectcommunity/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/selectcommunity/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.selectcommunity.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.selectcommunity.SelectCommunityMviModel
+
+expect fun getSelectCommunityViewModel(): SelectCommunityMviModel
