+++ settings.gradle.kts
@@ -46,3 +46,4 @@ include(":core-commonui:unit-ban")
 include(":core-commonui:unit-chat")
 include(":core-commonui:unit-zoomableimage")
 include(":core-commonui:unit-web")
+include(":core-commonui:unit-selectcommunity")
